Fixed formatting in README.
Just updated the README with some info for now. If this app is completed successfully and works very well, it could become useful to other FRC teams. If it comes to that point, we'll need good documentation in our code and this README to guide others through the process of modifying and using this Dashboard245 application.